+++ src/typescript/wastelandVentures.ts
@@ -1,5 +1,8 @@
 import WvActor from "./actor/wvActor.js";
 import WvActorSheet from "./actor/wvActorSheet.js";
+import { CONSTANTS } from "./constants.js";
+import { DragRuler } from "./integrations/dragRuler/dragRuler.js";
+import { createWvSpeedProvider } from "./integrations/dragRuler/wvSpeedProvider.js";
 import WvItem from "./item/wvItem.js";
 
 Hooks.once("init", () => {
@@ -12,3 +15,15 @@ Hooks.once("init", () => {
     makeDefault: true
   });
 });
+
+Hooks.once(
+  "dragRuler.ready",
+  (speedProvider: typeof DragRuler.SpeedProvider) => {
+    if (dragRuler) {
+      dragRuler.registerSystem(
+        CONSTANTS.systemId,
+        createWvSpeedProvider(speedProvider)
+      );
+    }
+  }
+);