+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- add support for the DragRuler foundry module, supporting ground movement and
+  sprinting highlights
+
 ## [0.0.2] - 2021-05-11
 
 ### Added